requirements.txt
# Production Dependencies
# Core AWS SDK (botocore managed by boto3)
boto3>=1.34.0,<2.0

# Image processing (for ragstack_common.image)
Pillow>=10.0.1,<11

# PDF processing (for ragstack_common.ocr)
PyMuPDF>=1.26.0,<2.0

# Web scraping (for ragstack_common.scraper)
httpx>=0.27.0
beautifulsoup4>=4.12.0
markdownify>=0.13.0
lxml>=5.0.0

# Development Dependencies
# Testing
pytest>=7.4.0,<9
pytest-cov>=4.1.0,<6.0
pytest-xdist>=3.3.1  # For parallel test execution
pytest-asyncio>=0.21.0

# Local Lambda testing
awslambdaric>=2.0.0

# Mocking AWS services
moto[s3,dynamodb,stepfunctions]>=5.0.0

# Code quality
ruff==0.14.2  # Pinned for consistent linting behavior across team/CI

# Text extraction (for ragstack_common.text_extractors)
ebooklib>=0.18      # EPUB extraction
python-docx>=1.1    # DOCX extraction
openpyxl>=3.1       # XLSX extraction
